#104 test: 멤버의 생활정보를 수정하고 상태코드 204를 리턴하는지 테스트
